Bit refactoring
