+++ src/main/java/io/stargate/sgv2/jsonapi/service/projection/DocumentProjector.java
@@ -6,12 +6,8 @@ import io.stargate.sgv2.jsonapi.exception.ErrorCode;
 import io.stargate.sgv2.jsonapi.exception.JsonApiException;
 import java.math.BigDecimal;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
-import java.util.regex.Pattern;
 
 /**
  * Helper class that implements functionality needed to support projections on documents fetched via
@@ -47,6 +43,10 @@ public class DocumentProjector {
 
   public static DocumentProjector identityProjector() {
     return IDENTITY_PROJECTOR;
+  }
+
+  public boolean isInclusion() {
+    return inclusion;
   }
 
   public void applyProjection(JsonNode document) {
@@ -189,110 +189,6 @@ public class DocumentProjector {
         ++inclusions;
         paths.add(path);
       }
-    }
-  }
-
-  /**
-   * Helper class that handles projection traversal for one level of nesting. Layers are either
-   * non-terminal (branches) or terminal (leaves)
-   */
-  private static class ProjectionLayer {
-    private static final Pattern DOT = Pattern.compile(Pattern.quote("."));
-
-    /** Whether this layer is terminal (matching) or branch (non-matching) */
-    private final boolean isTerminal;
-
-    /**
-     * Full path either to this layer (terminal), or to the first path through it (non-terminal) --
-     * needed for conflict reporting.
-     */
-    private String fullPath;
-
-    /** For non-terminal layers, segment-indexed next layers */
-    private final Map<String, ProjectionLayer> nextLayers;
-
-    ProjectionLayer(boolean terminal, String fullPath) {
-      isTerminal = terminal;
-      this.fullPath = fullPath;
-      nextLayers = isTerminal ? null : new HashMap<>();
-    }
-
-    public static ProjectionLayer buildLayers(Collection<String> dotPaths, boolean addDocId) {
-      // Root is always branch (not terminal):
-      ProjectionLayer root = new ProjectionLayer(false, "");
-      for (String fullPath : dotPaths) {
-        String[] segments = DOT.split(fullPath);
-        buildPath(fullPath, root, segments);
-      }
-      // May need to add doc-id inclusion/exclusion as well
-      if (addDocId) {
-        buildPath(
-            DocumentConstants.Fields.DOC_ID, root, new String[] {DocumentConstants.Fields.DOC_ID});
-      }
-      return root;
-    }
-
-    static void buildPath(String fullPath, ProjectionLayer layer, String[] segments) {
-      // First create branches
-      final int last = segments.length - 1;
-      for (int i = 0; i < last; ++i) {
-        // Try to find or create branch
-        layer = layer.findOrCreateBranch(fullPath, segments[i]);
-      }
-      // And then attach terminal (leaf)
-      layer.addTerminal(fullPath, segments[last]);
-    }
-
-    ProjectionLayer findOrCreateBranch(String fullPath, String segment) {
-      // Cannot proceed past terminal layer (shorter path):
-      if (isTerminal) {
-        reportPathConflict(this.fullPath, fullPath);
-      }
-      ProjectionLayer next = nextLayers.get(segment);
-      if (next == null) {
-        next = new ProjectionLayer(false, fullPath);
-        nextLayers.put(segment, next);
-      }
-      return next;
-    }
-
-    void addTerminal(String fullPath, String segment) {
-      // Cannot proceed past terminal layer (shorter path):
-      if (isTerminal) {
-        reportPathConflict(this.fullPath, fullPath);
-      }
-      // But will also not allow existing longer path:
-      ProjectionLayer next = nextLayers.get(segment);
-      if (next != null) {
-        reportPathConflict(fullPath, next.fullPath);
-      }
-      nextLayers.put(segment, new ProjectionLayer(true, fullPath));
-    }
-
-    void reportPathConflict(String fullPath1, String fullPath2) {
-      throw new JsonApiException(
-          ErrorCode.UNSUPPORTED_PROJECTION_PARAM,
-          ErrorCode.UNSUPPORTED_PROJECTION_PARAM.getMessage()
-              + ": projection path conflict between '"
-              + fullPath1
-              + "' and '"
-              + fullPath2
-              + "'");
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (o == this) return true;
-      if (!(o instanceof ProjectionLayer)) return false;
-      ProjectionLayer other = (ProjectionLayer) o;
-      return (this.isTerminal == other.isTerminal)
-          && Objects.equals(this.fullPath, other.fullPath)
-          && Objects.equals(this.nextLayers, other.nextLayers);
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(isTerminal, fullPath, nextLayers);
     }
   }
 }
